+++ addon/computed.js
@@ -1,0 +1,11 @@
+import Ember from 'ember';
+
+const {
+  computed
+} = Ember;
+
+export function transient(modelName, id) {
+  return computed(function() {
+    return this.transient(modelName, id);
+  }).readOnly();
+}