+++ addon/database/database-deserialize.js
@@ -15,6 +15,12 @@ export default Ember.Mixin.create({
     return internal;
   },
 
+  _createTransientInternalModel(modelClass, modelId, props) {
+    let internal = this.get('store')._createTransientInternalModel(modelClass, this, modelId, props);
+    this._storeSavedInternalModel(internal);
+    return internal;
+  },
+
   _deserializeInternalModelAttachments(internal, doc) {
     let definition = internal.definition;
     internal.withPropertyChanges(changed => {
@@ -31,6 +37,11 @@ export default Ember.Mixin.create({
     if(!internal) {
       return;
     }
+
+    assert({
+      error: 'transient',
+      reason: `cannot deserialize document '${docId}' delete for transient model '${internal.modelName}'`
+    }, !internal.transient);
 
     if(!internal.shouldDeserializeRevision(rev)) {
       return internal;
@@ -63,6 +74,11 @@ export default Ember.Mixin.create({
   _deserializeSavedDocumentToInternalModel(doc, expectedModelClass, optional=true, type=null) {
     let docId = doc._id;
     let rev = doc._rev;
+    let transient = !!doc._transient;
+
+    if(type !== 'shoebox') {
+      transient = false;
+    }
 
     let modelClass = this._modelClassForDocument(doc);
     if(!modelClass && !expectedModelClass) {
@@ -86,7 +102,18 @@ export default Ember.Mixin.create({
     } else {
       definition = this._definitionForModelClass(modelClass);
       let modelId = definition.modelId(docId);
-      internal = this._createExistingInternalModel(modelClass, modelId, false);
+      if(transient) {
+        internal = this._createTransientInternalModel(modelClass, modelId);
+      } else {
+        internal = this._createExistingInternalModel(modelClass, modelId);
+      }
+    }
+
+    if(!transient) {
+      assert({
+        error: 'transient',
+        reason: `cannot deserialize document '${docId}' for transient model '${internal.modelName}'`
+      }, !internal.transient);
     }
 
     if(internal.shouldDeserializeRevision(rev)) {
@@ -137,7 +164,7 @@ export default Ember.Mixin.create({
       let definition = internal.definition;
       assert({
         error: 'invalid_document',
-        reason: `document '${docId} is expected to be ${get(modelClass.class, 'modelName')} not ${definition.modelName}`
+        reason: `document '${docId}' is expected to be ${get(modelClass.class, 'modelName')} not ${definition.modelName}`
       }, definition.is(modelClass));
       return internal;
     }
@@ -157,11 +184,10 @@ export default Ember.Mixin.create({
     this._storeDeletedInternalModel(internal);
   },
 
-  _existingInternalModelForModelClass(modelClass, modelId, opts) {
-    let { create, deleted } = merge({ create: false, deleted: false }, opts);
+  _existingInternalModelForModelClass(modelClass, modelId, opts, props) {
+    let { create, deleted, transient } = merge({ create: false, deleted: false, transient: false }, opts);
 
     let definition = this._definitionForModelClass(modelClass);
-
     let docId = definition.docId(modelId);
 
     let internal = this._internalModelWithDocId(docId, deleted);
@@ -170,16 +196,36 @@ export default Ember.Mixin.create({
         internal = this._internalModelWithDocId(docId, true);
       }
       if(!internal) {
-        internal = this._createExistingInternalModel(modelClass, modelId);
+        if(transient) {
+          internal = this._createTransientInternalModel(modelClass, modelId, props);
+        } else {
+          internal = this._createExistingInternalModel(modelClass, modelId);
+        }
       }
+    }
+
+    if(transient) {
+      assert({
+        error: 'transient',
+        reason: `model '${definition.modelName}' with id '${modelId}' is already loaded`
+      }, internal.transient);
     }
 
     return internal;
   },
 
+  _transientInternalModelForModelClass(modelClass, modelId, props) {
+    return this._existingInternalModelForModelClass(modelClass, modelId, { create: true, transient: true }, props);
+  },
+
   _existingInternalModelForModelName(modelName, modelId, opts) {
     let modelClass = this.modelClassForName(modelName);
     return this._existingInternalModelForModelClass(modelClass, modelId, opts);
+  },
+
+  _transientInternalModelForModelName(modelName, modelId, props) {
+    let modelClass = this.modelClassForName(modelName);
+    return this._transientInternalModelForModelClass(modelClass, modelId, props);
   }
 
 });