+++ tests/dummy/app/services/store.js
@@ -2,19 +2,11 @@ import Ember from 'ember';
 import { Store } from 'sofa';
 
 const {
-  computed
+  computed: { reads }
 } = Ember;
 
-const root = () => {
-  return computed(function() {
-    return this.get('db.main').model('dummy-root', { id: 'singleton' });
-  }).readOnly();
-};
-
 const prop = key => {
-  return computed(function() {
-    return this.get('root').get(key);
-  }).readOnly();
+  return reads(`root.${key}`).readOnly();
 };
 
 export default Store.extend({
@@ -35,7 +27,7 @@ export default Store.extend({
     }
   },
 
-  root: root(),
+  root: reads('db.main.root'),
 
   allModels: prop('all'),
   dirtyModels: prop('dirty'),